fix(openfold3): preserve application isolation
Keep qualification matching local to its example consumer and remove the family test's reverse dependency across the application boundary.
